+++ app/Http/Controllers/InternshipCompanyController.php
@@ -10,29 +10,35 @@ class InternshipCompanyController extends Controller
     public function index()
     {
         $internship_company = InternshipCompany::all();
-        return response()->json(['message' => 'Success', 'data' => $internship_company]);
+        return response()->json(["message" => "Success", "data" => $internship_company]);
     }
 
     public function store(Request $request)
     {
         $data = $request->validate([
-            'internship_id' => 'required|integer',
-            'since' => 'required|string',
-            'sectors' => 'array',
-            'services' => 'array',
-            'address' => 'nullable|string',
-            'telephone' => 'nullable|string',
-            'email' => 'nullable|string',
-            'website' => 'nullable|string',
-            'director' => 'nullable|string',
-            'director_phone' => 'nullable|string',
-            'advisors' => 'nullable|string',
-            'structure' => 'nullable|string',
+            "internship_id" => "required|integer",
+            "since" => "required|string",
+            "sectors" => "array",
+            "services" => "array",
+            "address" => "nullable|string",
+            "telephone" => "nullable|string",
+            "email" => "nullable|string",
+            "website" => "nullable|string",
+            "director" => "nullable|string",
+            "director_phone" => "nullable|string",
+            "advisors" => "nullable|string",
+            "structure" => "nullable|string",
         ]);
 
-        $internship_company = InternshipCompany::create($data);
+        if (auth()->check()) {
+            auth()->user()->id;
+            $internship_company = InternshipCompany::create($data);
+    
+            return response()->json(["message" => "Company created successfully", "data" => $internship_company]);
+        } else {
+            return response()->json(["error" => true, "message" => "User not authenticated"]);
+        }
 
-        return response()->json(['message' => 'Company created successfully', 'data' => $internship_company]);
     }
 
     public function show($id)
@@ -40,34 +46,41 @@ class InternshipCompanyController extends Controller
         $internship_company = InternshipCompany::find($id);
 
         if (!$internship_company) {
-            return response()->json(['message' => 'Internship Company not found'], 404);
+            return response()->json(["message" => "Internship Company not found"], 404);
         }
-        return response()->json(['message' => 'Success', 'data' => $internship_company]);
+        return response()->json(["message" => "Success", "data" => $internship_company]);
     }
 
     public function update(Request $request, $id)
     {
         $data = $request->validate([
-            'internship_id' => 'integer',
-            'since' => 'string',
-            'sectors' => 'array',
-            'services' => 'array',
-            'address' => 'nullable|string',
-            'telephone' => 'nullable|string',
-            'email' => 'nullable|string',
-            'website' => 'nullable|string',
-            'director' => 'nullable|string',
-            'director_phone' => 'nullable|string',
-            'advisors' => 'nullable|string',
-            'structure' => 'nullable|string',
+            "internship_id" => "integer",
+            "since" => "string",
+            "sectors" => "array",
+            "services" => "array",
+            "address" => "nullable|string",
+            "telephone" => "nullable|string",
+            "email" => "nullable|string",
+            "website" => "nullable|string",
+            "director" => "nullable|string",
+            "director_phone" => "nullable|string",
+            "advisors" => "nullable|string",
+            "structure" => "nullable|string",
         ]);
 
         $internship_company = InternshipCompany::find($id);
+        
         if (!$internship_company) {
-            return response()->json(['message' => 'Internship Company not found'], 404);
+            return response()->json(["message" => "Internship Company not found"], 404);
         }
-        $internship_company->update($data);
-        return response()->json(['message' => 'Internship Company updated successfully', 'data' => $internship_company]);
+
+        if (auth()->check()) {
+            auth()->user()->id;
+            $internship_company->update($data);
+            return response()->json(["message" => "Internship Company updated successfully", "data" => $internship_company]);
+        } else {
+            return response()->json(["error" => true, "message" => "User not authenticated"]);
+        }
     }
 
     public function destroy($id)
@@ -75,11 +88,17 @@ class InternshipCompanyController extends Controller
         $internship_company = InternshipCompany::find($id);
 
         if (!$internship_company) {
-            return response()->json(['message' => 'Internship Company not found'], 404);
+            return response()->json(["message" => "Internship Company not found"], 404);
         }
 
-        $internship_company->delete();
+        if (auth()->check()) {
+            auth()->user()->id;
+            $internship_company->delete();
+    
+            return response()->json(["message" => "Internship Company deleted successfully"]);
+        } else {
+            return response()->json(["error" => true, "message" => "User not authenticated"]);
+        }
 
-        return response()->json(['message' => 'Internship Company deleted successfully']);
     }
 }
